+++ hutool-core/src/test/java/cn/hutool/core/lang/ClassScanerTest.java
@@ -10,7 +10,7 @@ public class ClassScanerTest {
 	@Test
 	@Ignore
 	public void scanTest() {
-		ClassScaner scaner = new ClassScaner("cn.hutool.core.util", null);
+		ClassScanner scaner = new ClassScanner("cn.hutool.core.util", null);
 		Set<Class<?>> set = scaner.scan();
 		for (Class<?> clazz : set) {
 			Console.log(clazz.getName());